Backend/requirements.txt
fastapi
uvicorn[standard]

passlib[bcrypt]
python-jose[cryptography]

python-dotenv

google-api-python-client
google-auth
google-auth-oauthlib
google-auth-httplib2